Removed special abilities bar & added obtain ability script
- Removed dependencies on special abilities bar script.
- Added script to allow destroying and obtaining of correct power-up.
- Fixed spawn count for power-ups so maxed spawned is 2 but more can
spawn once there is less than 2 in the scene.

+++ Assets/Scripts/PillarOfLight/PillarHealth.cs
@@ -5,13 +5,13 @@ namespace PillarOfLight
     public class PillarHealth : MonoBehaviour
     {
         
-        public float startingHealth = 100;
+        public int startingHealth = 100;
         public Material pillarOfLightMaterial;
         public GameObject exitSign;
         public GameObject replaySign;
         public GameObject pillarExplosionEffect;
         
-        private float _currentHealth;
+        private int _currentHealth;
         private float _pillarColor = 255f;
         private int _hitCount = 0;
         private Material _pillarOfLightMaterialInstance;
@@ -69,6 +69,6 @@ namespace PillarOfLight
         }
         
         // Calculating random power-up drops
-        public float CurrentHealth => _currentHealth;
+        public int CurrentHealth => _currentHealth;
     }
 }
